Separate element for plugin specific settings.
